+++ src/styles/layout/header/header.module.css
@@ -1,0 +1,40 @@
+/* 헤더 틀 */
+.container {
+	width: 100vw;
+	min-height: 120px;
+
+	display: flex;
+	justify-content: center;
+
+	background-color: #f4f2e5;
+}
+.wrapper {
+	width: 1300px;
+	min-height: inherit;
+
+	display: flex;
+	flex-direction: column;
+}
+
+/* 헤더 상단 */
+.topWrapper {
+	min-height: 60px;
+
+	display: flex;
+	align-items: center;
+	justify-content: space-between;
+}
+.logo {
+	width: 250px;
+	height: 60px;
+
+	background-color: yellow;
+}
+
+/* 헤더 하단 */
+.bottomWrapper {
+	min-height: 60px;
+
+	display: flex;
+	justify-content: space-between;
+}
