+++ src/styles/layout/header/nav.module.css
@@ -1,0 +1,21 @@
+.navigationWrapper {
+	height: inherit;
+
+	display: flex;
+	align-items: center;
+}
+.navigationWrapper > a {
+	width: inherit;
+	height: 20px;
+	font-size: 16px;
+	font-weight: bold;
+	margin-right: 25px;
+
+	color: #000000;
+	text-decoration-line: none;
+
+	transition: color 0.2s;
+}
+.navigationWrapper > a:hover {
+	color: #6594e7;
+}
