+++ src/styles/layout/header/search.module.css
@@ -1,0 +1,31 @@
+.searchForm {
+	width: 400px;
+	height: 60px;
+
+	display: flex;
+	align-items: center;
+	position: relative;
+}
+.searchForm > input {
+	width: 98%;
+	height: 60%;
+	border-radius: 50px;
+	padding: 0 20px;
+
+	border: none;
+	outline: none;
+
+	transition: border-radius 0.3s;
+}
+.searchForm > input:focus {
+	border-radius: 5px;
+}
+.searchIcon {
+	right: 15px;
+
+	border: none;
+	position: absolute;
+	cursor: pointer;
+
+	background-color: #ffffff;
+}
